==> 03_Regresiones_viejo.qmd <==
---
title: "1er Trabajo Práctico de Econometría"
author:
  - name: Mauro Ciani
  - name: Juan Camilo Gutman

abstract: > 
  El presente trabajo se propuso blabla
---

## Librerías

```{r}
#| output: false

library(tidyverse)    #Para manejar bases de datos
library(ggplot2)      #Para graficar
library(modelsummary) #Mejores tablas de regresión 
library(tinytable)    #Motor de creación de tablas
```

```{r}
#| include: false

options(tinytable_tt_digits = 3)
options(tinytable_theme_placement_latex_float = "H")
```

```{r}
df1 <- readRDS("bases/eph_1abc.RDS")
# df2 <- readRDS("Bases/eph_1de.RDS")
```

Nos quedamos únicamente con los jefes de hogar, hombres, con edades entre 25 y 65 años, ocupados y asalariados. Habría que mencionar algo sobre la base resultante (y cuantos datos estamos descartando) y un mínimo análisis exploratorio.

## Probando referencias

[@pradier2023] [@arel-bundock2022]

Cras a molestie nulla. [@Hemingway1952]

Siguiendo a [@koenker2005], encontramos que:

```{r}
#| echo: true
#| file: 01B_delta.R
```

[@pradier2023]

En la tabla pueden leerse los resultados de la primer regresión.

```{=tex}
\newpage
\KOMAoptions{paper=landscape,pagesize}
\recalctypearea
```
```{r}
#| echo: false

reg1 <- lm(logSal ~ edad + educf + est_civ + region,
           data = df1
           )
reg2 <- lm(logSal ~ edad + educn + est_civ + region,
           data = df1
           )
regs <- list("Sin edad^2" = reg1,
             "Con edad^2" = reg2
             )

#coefmap
cm1y2 <- c("(Intercept)"="Constante",
        "Num.Obs." = "n")

notas1y2 <- c("Notes: Aca irían nuestras notas. Comentarios, explicaciones. Elegancia.")

epigrafe1y2 <- 'Tabla 1 — Resultados de las regresiones, Ecuación de Mincer
(estimación del efecto de la educación en el logaritmo de los salarios, controlando por...)'
# 
# rdos_regs1y2 <- modelsummary(regs,
#                            escape = TRUE,
#                            shape = term ~ model + statistic,
#                            cap = "1eras regresiones",
#                            #width = 0.8,
#                            # estimate = "estimate",
#                            coef_rename = coef_rename,
#                            #coef_map = cm,
#                            estimate="{estimate}{stars}",
#                            statistic = c("p.value",
#                                          "conf.low",
#                                          "conf.high"
#                                          ),
#                            stars = c('*' = .1,
#                                      '**' = .05,
#                                      '***'=0.01
#                                      ),
#                            notes = notas1y2,
#                            title = epigrafe1y2
#                            )

# rdos_regs1y2 %>%
#   theme_tt("multipage",
#            rowhead = 1
#            ) %>%
#   group_tt(i = list("Nivel Educativo" = 3,
#                     "Estado Civil" = 9,
#                     "Aglomerado" = 13,
#                     "Región" = 44
#                     ),
#            )
```

```{r}
#| echo: false

#coefmap
cm2 <- c("(Intercept)"="Constante",
        "Num.Obs." = "n")
#Notas
notas2 <- c("Notas: Acá comentarios, explicaciones, etc.",
            "***Significant at the 1 percent level.",
            "**Significant at the 5 percent level.",
            "*Significant at the 10 percent level."
            )

epigrafe2 <- 'Tabla 2 — Resultados de la regresión, Ecuación de Mincer
(estimación del efecto de la educación en el logaritmo de los salarios, controlando por...)'
rdo_reg2 <- modelsummary(reg2,
                           escape = TRUE,
                           shape = term ~ model + statistic,
                           cap = "1eras regresión",
                           #width = 0.8,
                           # estimate = "estimate",
                           coef_rename = coef_rename,
                           #coef_map = cm,
                           estimate="{estimate}{stars}",
                           statistic = c("p.value",
                                         "conf.low",
                                         "conf.high"
                                         ),
                           stars = c('*' = .1,
                                     '**' = .05,
                                     '***'=0.01
                                     ),
                           notes = notas2,
                           title = epigrafe2
                           )
rdo_reg2 %>%
  theme_tt("multipage",
           rowhead = 1
           ) %>%
  group_tt(i = list("Nivel Educativo" = 3,
                    "Estado Civil" = 9,
                    "Región" = 13,
                    "Bondad de Ajuste" = 18
                    ),
           )
```

```{r}
gof2 <- get_gof(reg2)
gof2 <- as.data.frame(t(gof2), optional = TRUE)
gof2$estad <- rownames(gof2)
gof2 <- gof2 %>% select(estad, everything())
names(gof2) <- NULL

gof1 <- get_gof(reg2)
gof1 <- as.data.frame(t(gof1), optional = TRUE)
gof1$estad <- rownames(gof1)
gof1 <- gof1 %>% select(estad, everything())
names(gof1) <- NULL
```

```{r}
#| layout-ncol: 2
# gof1 <- gof1[4:nrow(gof1)-1,]
# 
# gof1$orden <- c(3, 1, 2, 4, 5)
# 
# gof1<- gof1 %>%
#   arrange(orden)
# 
# gof1$orden <- NULL
# 
# gof2 %>% tt()
```

```{r}
gm1 <- gof_map
gm1[1,2] <- "n"
gm1[16:19,4] <- "TRUE" #Chau a los AIC y BIC
gm1[16:19,]

gof1 <- get_gof(reg1, gof_map = gm1)
gof1 %>% tt()
```

```{r}
# Anova(reg2) %>% tt()
```

\input

testlineal.tex

```{=tex}
\newpage
\KOMAoptions{paper=portrait,pagesize}
\recalctypearea
```
texto de prueba

# Bibliografía
